refactor: handling exception to id false

+++ forum/src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -8,7 +8,6 @@ import br.com.alura.forum.modelo.Topico;
 import br.com.alura.forum.repository.TopicoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.util.UriComponentsBuilder;
 
@@ -16,6 +15,7 @@ import javax.transaction.Transactional;
 import javax.validation.Valid;
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 
 @RestController
@@ -50,11 +50,15 @@ public class TopicosController {
     }
 
     @GetMapping("/{id}")
-    public DetalhesTopicoDTO detalhar(@PathVariable Long id){
-        Topico topico = topicoRepository.getOne(id);
+    public ResponseEntity<DetalhesTopicoDTO> detalhar(@PathVariable Long id){
+        Optional<Topico> topico = topicoRepository.findById(id);
+
+        if(topico.isPresent()){
+            return ResponseEntity.ok(new DetalhesTopicoDTO(topico.get()));
+        }
 
 
-        return new DetalhesTopicoDTO(topico);
+        return ResponseEntity.notFound().build();
     }
 
     /*
@@ -65,16 +69,28 @@ public class TopicosController {
     @PutMapping("/{id}")
     @Transactional
     public ResponseEntity<TopicoDTO> atualizar (@PathVariable Long id,@RequestBody @Valid AtualizarTopicoForm form ){
-        Topico topico = form.atualizar(id, topicoRepository);
+        Optional<Topico> optional = topicoRepository.findById(id);
 
-        return ResponseEntity.ok(new TopicoDTO(topico));
+        if(optional.isPresent()){
+            Topico topico = form.atualizar(id, topicoRepository);
+            return ResponseEntity.ok(new TopicoDTO(topico));
+        }
+
+        return ResponseEntity.notFound().build();
     }
 
     @DeleteMapping("/{id}")
     @Transactional
     public ResponseEntity<?> deletar(@PathVariable Long id){
-        topicoRepository.deleteById(id);
 
-        return ResponseEntity.ok().build();
+        Optional<Topico> optional = topicoRepository.findById(id);
+
+        if(optional.isPresent()){
+            topicoRepository.deleteById(id);
+            return ResponseEntity.ok().build();
+        }
+
+        return ResponseEntity.notFound().build();
+
     }
 }
